pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cascade"
version = "0.3.0"
description = "AI-powered research intelligence — web-native"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "rich>=13.0",
    "openai>=1.0",
    "anthropic>=0.30",
    "arxiv>=2.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pymupdf>=1.24",
    "pyyaml>=6.0",
    "tiktoken>=0.6",
    "chromadb>=0.5",
    "pyvis>=0.3",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=4.0", "ruff>=0.3"]

[project.scripts]
cascade = "cascade.server:main"

[tool.setuptools.packages.find]
include = ["cascade*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
